requirements.txt
blinker
click
colorama
dominate
Flask-Bootstrap
Flask-Login
flask-sqlalchemy
Flask-WTF
Flask
greenlet
gunicorn==20.1.0
itsdangerous
Jinja2
MarkupSafe
psycopg2-binary
python-dotenv
setuptools
SQLAlchemy
typing-extensions
visitor
Werkzeug
wheel
WTForms
